+++ common/src/main/java/generations/gg/generations/structures/generationsstructures/tags/GenerationsBiomeTags.java
@@ -51,6 +51,9 @@ public class GenerationsBiomeTags {
     /** the tag for biomes that can have a Tapu shrine */
     public static final TagKey<Biome> HAS_TAPU_SHRINE = create("has_structure/tapu_shrine");
 
+    /** the tag for biomes that can have a haunted mansion */
+    public static final TagKey<Biome> HAS_HAUNTED_MANSION = create("has_structure/haunted_mansion");
+
     /**
      * Creates a new {@link TagKey} for the given name.
      * @param name The name of the tag.