+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.1.3-Beta
+- Add Haunted Mansion Shrine
 
 # 1.1.2-Beta
 - Fix Meowth Balloon having sign and no loot